Experiment with an actor pool and nonblocking receive on Fetcher
In theory this should cut down on the number of threads, though
we still end up with a lot more than truly seems necessary.
(Configuring AsyncHttpClient to allow fewer might be good,
depending on how it works internally.)

With each actor in the pool handling one outstanding request
at a time, the number of actors in the pool limits how
many requests we make at a time. Also, fetcher actors could
store state related to their current request, if they wanted
to - that might be handy.